requirements.txt
pyahocorasick==2.2.0
Flask==3.1.2
langchain_chroma==1.0.0
langchain_core==1.0.4
langchain_openai==1.0.2
langgraph==1.0.3
psutil==7.1.3
pydantic==2.12.4
pydantic_settings==2.12.0
python-dotenv==1.2.1
redis==7.0.1
Requests==2.32.5
sentence_transformers==5.1.2
